+++ src/main/java/com/example/whynotpc/persistence/jwt/RefreshTokenRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RefreshTokenRepo extends JpaRepository<RefreshToken, Integer> {
+public interface RefreshTokenRepo extends JpaRepository<RefreshToken, Long> {
     Optional<RefreshToken> findByToken(String token);
 }
